做了一些修改 多线程运行
1. CountLine判断空行的规则改变了，修复了行数读取错误 isValidate弃用
2. CountWord类弃用，合法单词的长度如果超过一定长度无法通过size方法获取。改为在SplitWord中 进行计算。
3. 修改了读取文件的一个错误，这个错误导致读取的文本存在问题。
4. 将单词数量计算和行数计算转为多线程运行，缩短了一定的总运行时间。

+++ 221801310/src/CountAsciiChar.java
@@ -2,16 +2,9 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 public class CountAsciiChar{
+
     /*统计文件ascii字符数量 传入参数text为文本内容字符串*/
-    public static long countChar(String text){
-//        int num = text.length();
-//
-//        String regex = "\\p{ASCII}";
-//        Pattern pattern = Pattern.compile(regex);
-//        Matcher matcher = pattern.matcher(text);
-//        while (matcher.find()){
-//            num++;
-//        }
+    public static int countChar(String text){
         return text.length();
     }
